Extract deprecated router views
