+++ app/api/api/views.py
@@ -13,23 +13,16 @@ from api.paginations import CustomPagination
 from api.serializers import (
     ConceptSerializer,
     GetRulesAnalysis,
-    MappingRuleSerializer,
-    OmopFieldSerializer,
-    OmopTableSerializer,
     ScanReportConceptSerializer,
     ScanReportCreateSerializer,
     ScanReportEditSerializer,
     ScanReportFieldEditSerializer,
-    ScanReportFieldListSerializer,
     ScanReportFieldListSerializerV2,
     ScanReportFilesSerializer,
     ScanReportTableEditSerializer,
-    ScanReportTableListSerializer,
     ScanReportTableListSerializerV2,
-    ScanReportValueEditSerializer,
     ScanReportValueViewSerializer,
     ScanReportValueViewSerializerV2,
-    ScanReportViewSerializer,
     ScanReportViewSerializerV2,
     UserSerializer,
 )
@@ -39,8 +32,7 @@ from datasets.serializers import DataPartnerSerializer
 from django.contrib.auth.models import User
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ObjectDoesNotExist
-from django.db.models.query_utils import Q
-from django.http import HttpResponse, JsonResponse
+from django.http import HttpResponse
 from django.utils.decorators import method_decorator
 from django.views.decorators.cache import cache_page
 from django.views.decorators.vary import vary_on_cookie
@@ -58,13 +50,11 @@ from shared.mapping.models import (
     DataPartner,
     MappingRule,
     OmopField,
-    OmopTable,
     ScanReport,
     ScanReportConcept,
     ScanReportField,
     ScanReportTable,
     ScanReportValue,
-    VisibilityChoices,
 )
 from shared.mapping.permissions import (
     CanAdmin,
@@ -84,22 +74,6 @@ from shared.services.rules_export import (
     get_mapping_rules_list,
     make_dag,
 )
-
-
-class ConceptViewSet(viewsets.ReadOnlyModelViewSet):
-    queryset = Concept.objects.all()
-    serializer_class = ConceptSerializer
-
-
-class ConceptFilterViewSet(viewsets.ReadOnlyModelViewSet):
-    queryset = Concept.objects.all()
-    serializer_class = ConceptSerializer
-    filter_backends = [DjangoFilterBackend]
-    filterset_fields = {
-        "concept_id": ["in", "exact"],
-        "concept_code": ["in", "exact"],
-        "vocabulary_id": ["in", "exact"],
-    }
 
 
 class ConceptFilterViewSetV2(viewsets.ReadOnlyModelViewSet):
@@ -132,136 +106,6 @@ class UserDetailView(APIView):
     def get(self, request, *args, **kwargs):
         serializer = UserSerializer(request.user)
         return Response(serializer.data)
-
-
-class ScanReportListViewSet(viewsets.ModelViewSet):
-    filter_backends = [DjangoFilterBackend]
-    filterset_fields = {"parent_dataset": ["exact"]}
-
-    def get_permissions(self):
-        if self.request.method == "DELETE":
-            # user must be able to view and be an admin to delete a scan report
-            self.permission_classes = [IsAuthenticated & CanView & CanAdmin]
-        elif self.request.method in ["PUT", "PATCH"]:
-            # user must be able to view and be either an editor or and admin
-            # to edit a scan report
-            self.permission_classes = [IsAuthenticated & CanView & (CanEdit | CanAdmin)]
-        else:
-            self.permission_classes = [IsAuthenticated & (CanView | CanEdit | CanAdmin)]
-        return [permission() for permission in self.permission_classes]
-
-    def get_serializer_class(self):
-        if self.request.method in ["GET", "POST"]:
-            # use the view serialiser if on GET requests
-            return ScanReportViewSerializer
-        if self.request.method in ["PUT", "PATCH", "DELETE"]:
-            # use the edit serialiser when the user tries to alter the scan report
-            return ScanReportEditSerializer
-        return super().get_serializer_class()
-
-    def get_queryset(self):
-        """
-        If the User is the `AZ_FUNCTION_USER`, return all ScanReports.
-
-        Else, apply the correct rules regarding the visibility of the Dataset and SR,
-        and the membership of the User of viewer/editor/admin/author for either.
-        """
-        if self.request.user.username == os.getenv("AZ_FUNCTION_USER"):
-            return ScanReport.objects.all().distinct()
-
-        return ScanReport.objects.filter(
-            (
-                Q(parent_dataset__visibility=VisibilityChoices.PUBLIC)
-                & (
-                    Q(
-                        # Dataset and SR are 'PUBLIC'
-                        visibility=VisibilityChoices.PUBLIC,
-                    )
-                    | (
-                        Q(visibility=VisibilityChoices.RESTRICTED)
-                        & (
-                            Q(
-                                # Dataset is public
-                                # SR is restricted but user is in SR viewers
-                                viewers=self.request.user.id,
-                            )
-                            | Q(
-                                # Dataset is public
-                                # SR is restricted but user is in SR editors
-                                editors=self.request.user.id,
-                            )
-                            | Q(
-                                # Dataset is public
-                                # SR is restricted but user is SR author
-                                author=self.request.user.id,
-                            )
-                            | Q(
-                                # Dataset is public
-                                # SR is restricted but user is in Dataset editors
-                                parent_dataset__editors=self.request.user.id,
-                            )
-                            | Q(
-                                # Dataset is public
-                                # SR is restricted but user is in Dataset admins
-                                parent_dataset__admins=self.request.user.id,
-                            )
-                        )
-                    )
-                )
-            )
-            | (
-                Q(parent_dataset__visibility=VisibilityChoices.RESTRICTED)
-                & (
-                    Q(
-                        # Dataset and SR are restricted
-                        # User is in Dataset admins
-                        parent_dataset__admins=self.request.user.id,
-                    )
-                    | Q(
-                        # Dataset and SR are restricted
-                        # User is in Dataset editors
-                        parent_dataset__editors=self.request.user.id,
-                    )
-                    | (
-                        Q(parent_dataset__viewers=self.request.user.id)
-                        & (
-                            Q(
-                                # Dataset and SR are restricted
-                                # User is in Dataset viewers and SR viewers
-                                viewers=self.request.user.id,
-                            )
-                            | Q(
-                                # Dataset and SR are restricted
-                                # User is in Dataset viewers and SR editors
-                                editors=self.request.user.id,
-                            )
-                            | Q(
-                                # Dataset and SR are restricted
-                                # User is in Dataset viewers and SR author
-                                author=self.request.user.id,
-                            )
-                            | Q(
-                                # Dataset is restricted
-                                # But SR is 'PUBLIC'
-                                visibility=VisibilityChoices.PUBLIC,
-                            )
-                        )
-                    )
-                )
-            ),
-            parent_dataset__project__members=self.request.user.id,
-        ).distinct()
-
-    def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(
-            data=request.data, many=isinstance(request.data, list)
-        )
-        serializer.is_valid(raise_exception=True)
-        self.perform_create(serializer)
-        headers = self.get_success_headers(serializer.data)
-        return Response(
-            serializer.data, status=status.HTTP_201_CREATED, headers=headers
-        )
 
 
 class ScanReportListViewSetV2(viewsets.ModelViewSet):
@@ -509,102 +353,6 @@ class StructuralMappingTableAPIView(APIView):
         return qs
 
 
-class ScanReportTableViewSet(viewsets.ModelViewSet):
-    queryset = ScanReportTable.objects.all()
-    filter_backends = [DjangoFilterBackend, OrderingFilter, ScanReportAccessFilter]
-    ordering_fields = ["name", "person_id", "event_date"]
-    filterset_fields = {
-        "scan_report": ["in", "exact"],
-        "name": ["in", "icontains"],
-        "id": ["in", "exact"],
-    }
-
-    ordering = "-created_at"
-
-    def get_permissions(self):
-        if self.request.method == "DELETE":
-            # user must be able to view and be an admin to delete a scan report
-            self.permission_classes = [IsAuthenticated & CanView & CanAdmin]
-        elif self.request.method in ["PUT", "PATCH"]:
-            # user must be able to view and be either an editor or and admin
-            # to edit a scan report
-            self.permission_classes = [IsAuthenticated & CanView & (CanEdit | CanAdmin)]
-        else:
-            self.permission_classes = [IsAuthenticated & (CanView | CanEdit | CanAdmin)]
-        return [permission() for permission in self.permission_classes]
-
-    def get_serializer_class(self):
-        if self.request.method in ["GET", "POST"]:
-            # use the view serialiser if on GET requests
-            return ScanReportTableListSerializer
-        if self.request.method in ["PUT", "PATCH", "DELETE"]:
-            # use the edit serialiser when the user tries to alter the scan report
-            return ScanReportTableEditSerializer
-        return super().get_serializer_class()
-
-    def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(
-            data=request.data, many=isinstance(request.data, list)
-        )
-        serializer.is_valid(raise_exception=True)
-        self.perform_create(serializer)
-        headers = self.get_success_headers(serializer.data)
-        return Response(
-            serializer.data, status=status.HTTP_201_CREATED, headers=headers
-        )
-
-    def partial_update(self, request: Any, *args: Any, **kwargs: Any) -> Response:
-        """
-        Perform a partial update on the instance.
-
-        Args:
-            request (Any): The request object.
-            *args (Any): Additional positional arguments.
-            **kwargs (Any): Additional keyword arguments.
-
-        Returns:
-            Response: The response object.
-        """
-        instance = self.get_object()
-        partial = kwargs.pop("partial", True)
-        serializer = self.get_serializer(instance, data=request.data, partial=partial)
-        serializer.is_valid(raise_exception=True)
-        self.perform_update(serializer)
-
-        # Delete the current mapping rules
-        delete_mapping_rules(instance.id)
-
-        # Map the table
-        scan_report_instance = instance.scan_report
-        data_dictionary_name = (
-            scan_report_instance.data_dictionary.name
-            if scan_report_instance.data_dictionary
-            else None
-        )
-
-        # Send to functions
-        msg = {
-            "scan_report_id": scan_report_instance.id,
-            "table_id": instance.id,
-            "data_dictionary_blob": data_dictionary_name,
-        }
-        base_url = f"{settings.AZ_URL}"
-        trigger = (
-            f"/api/orchestrators/{settings.AZ_RULES_NAME}?code={settings.AZ_RULES_KEY}"
-        )
-        try:
-            response = requests.post(urljoin(base_url, trigger), json=msg)
-            response.raise_for_status()
-        except request.exceptions.HTTPError as e:
-            logging.error(f"HTTP Trigger failed: {e}")
-
-        # TODO: The worker_id can be used for status, but we need to save it somewhere.
-        # resp_json = response.json()
-        # worker_id = resp_json.get("instanceId")
-
-        return Response(serializer.data)
-
-
 class ScanReportTableViewSetV2(viewsets.ModelViewSet):
     queryset = ScanReportTable.objects.all()
     filterset_fields = {
@@ -701,47 +449,6 @@ class ScanReportTableViewSetV2(viewsets.ModelViewSet):
         return Response(serializer.data)
 
 
-class ScanReportFieldViewSet(viewsets.ModelViewSet):
-    queryset = ScanReportField.objects.all()
-    filter_backends = [DjangoFilterBackend, ScanReportAccessFilter]
-    filterset_fields = {
-        "scan_report_table": ["in", "exact"],
-        "name": ["in", "exact"],
-    }
-
-    def get_permissions(self):
-        if self.request.method == "DELETE":
-            # user must be able to view and be an admin to delete a scan report
-            self.permission_classes = [IsAuthenticated & CanView & CanAdmin]
-        elif self.request.method in ["PUT", "PATCH"]:
-            # user must be able to view and be either an editor or and admin
-            # to edit a scan report
-            self.permission_classes = [IsAuthenticated & CanView & (CanEdit | CanAdmin)]
-        else:
-            self.permission_classes = [IsAuthenticated & CanView | CanEdit | CanAdmin]
-        return [permission() for permission in self.permission_classes]
-
-    def get_serializer_class(self):
-        if self.request.method in ["GET", "POST"]:
-            # use the view serialiser if on GET requests
-            return ScanReportFieldListSerializer
-        if self.request.method in ["PUT", "PATCH", "DELETE"]:
-            # use the edit serialiser when the user tries to alter the scan report
-            return ScanReportFieldEditSerializer
-        return super().get_serializer_class()
-
-    def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(
-            data=request.data, many=isinstance(request.data, list)
-        )
-        serializer.is_valid(raise_exception=True)
-        self.perform_create(serializer)
-        headers = self.get_success_headers(serializer.data)
-        return Response(
-            serializer.data, status=status.HTTP_201_CREATED, headers=headers
-        )
-
-
 class ScanReportFieldViewSetV2(viewsets.ModelViewSet):
     # Need to reset to see the change in data returned?!
     queryset = ScanReportField.objects.all()
@@ -779,62 +486,6 @@ class ScanReportFieldViewSetV2(viewsets.ModelViewSet):
     @method_decorator(vary_on_cookie)
     def list(self, request, *args, **kwargs):
         return super().list(request, *args, **kwargs)
-
-
-class ScanReportConceptViewSet(viewsets.ModelViewSet):
-    queryset = ScanReportConcept.objects.all()
-    serializer_class = ScanReportConceptSerializer
-
-    def create(self, request, *args, **kwargs):
-        body = request.data
-        if not isinstance(body, list):
-            # Extract the content_type
-            content_type_str = body.pop("content_type", None)
-            content_type = ContentType.objects.get(model=content_type_str)
-            body["content_type"] = content_type.id
-
-            concept = ScanReportConcept.objects.filter(
-                concept=body["concept"],
-                object_id=body["object_id"],
-                content_type=content_type,
-            )
-            if concept.count() > 0:
-                print("Can't add multiple concepts of the same id to the same object")
-                response = JsonResponse(
-                    {
-                        "status_code": 403,
-                        "ok": False,
-                        "statusText": "Can't add multiple concepts of the same id to the same object",
-                    }
-                )
-                response.status_code = 403
-                return response
-        else:
-            # for each item in the list, identify any existing SRConcepts that clash, and block their creation
-            # this method may be quite slow as it has to wait for each query
-            filtered = []
-            for item in body:
-                # Extract the content_type
-                content_type_str = item.pop("content_type", None)
-                content_type = ContentType.objects.get(model=content_type_str)
-                item["content_type"] = content_type.id
-
-                concept = ScanReportConcept.objects.filter(
-                    concept=item["concept"],
-                    object_id=item["object_id"],
-                    content_type=content_type,
-                )
-                if concept.count() == 0:
-                    filtered.append(item)
-            body = filtered
-
-        serializer = self.get_serializer(data=body, many=isinstance(body, list))
-        serializer.is_valid(raise_exception=True)
-        self.perform_create(serializer)
-        headers = self.get_success_headers(serializer.data)
-        return Response(
-            serializer.data, status=status.HTTP_201_CREATED, headers=headers
-        )
 
 
 class ScanReportConceptViewSetV2(viewsets.ModelViewSet):
@@ -931,18 +582,6 @@ class ScanReportConceptViewSetV2(viewsets.ModelViewSet):
         )
 
 
-class ScanReportConceptFilterViewSet(viewsets.ModelViewSet):
-    queryset = ScanReportConcept.objects.all()
-    serializer_class = ScanReportConceptSerializer
-    filter_backends = [DjangoFilterBackend]
-    filterset_fields = {
-        "concept__concept_id": ["in", "exact"],
-        "object_id": ["in", "exact"],
-        "id": ["in", "exact"],
-        "content_type": ["in", "exact"],
-    }
-
-
 class ScanReportConceptFilterViewSetV2(viewsets.ModelViewSet):
     queryset = ScanReportConcept.objects.all().order_by("id")
     serializer_class = ScanReportConceptSerializer
@@ -970,21 +609,6 @@ class DataPartnerViewSet(viewsets.ModelViewSet):
         return Response(
             serializer.data, status=status.HTTP_201_CREATED, headers=headers
         )
-
-
-class OmopTableViewSet(viewsets.ModelViewSet):
-    queryset = OmopTable.objects.all()
-    serializer_class = OmopTableSerializer
-
-
-class OmopFieldViewSet(viewsets.ModelViewSet):
-    queryset = OmopField.objects.all()
-    serializer_class = OmopFieldSerializer
-
-
-class MappingRuleViewSet(viewsets.ModelViewSet):
-    queryset = MappingRule.objects.all()
-    serializer_class = MappingRuleSerializer
 
 
 class RulesList(viewsets.ModelViewSet):
@@ -1111,49 +735,6 @@ class AnalyseRules(viewsets.ModelViewSet):
     filterset_fields = ["id"]
 
 
-class ScanReportValueViewSet(viewsets.ModelViewSet):
-    queryset = ScanReportValue.objects.all()
-    filter_backends = [DjangoFilterBackend, ScanReportAccessFilter]
-    filterset_fields = {
-        "scan_report_field": ["in", "exact"],
-        "value": ["in", "exact"],
-        "id": ["in", "exact"],
-    }
-    ordering = "id"
-
-    def get_permissions(self):
-        if self.request.method == "DELETE":
-            # user must be able to view and be an admin to delete a scan report
-            self.permission_classes = [IsAuthenticated & CanView & CanAdmin]
-        elif self.request.method in ["PUT", "PATCH"]:
-            # user must be able to view and be either an editor or and admin
-            # to edit a scan report
-            self.permission_classes = [IsAuthenticated & CanView & (CanEdit | CanAdmin)]
-        else:
-            self.permission_classes = [IsAuthenticated & (CanView | CanEdit | CanAdmin)]
-        return [permission() for permission in self.permission_classes]
-
-    def get_serializer_class(self):
-        if self.request.method in ["GET", "POST"]:
-            # use the view serialiser if on GET requests
-            return ScanReportValueViewSerializer
-        if self.request.method in ["PUT", "PATCH", "DELETE"]:
-            # use the edit serialiser when the user tries to alter the scan report
-            return ScanReportValueEditSerializer
-        return super().get_serializer_class()
-
-    def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(
-            data=request.data, many=isinstance(request.data, list)
-        )
-        serializer.is_valid(raise_exception=True)
-        self.perform_create(serializer)
-        headers = self.get_success_headers(serializer.data)
-        return Response(
-            serializer.data, status=status.HTTP_201_CREATED, headers=headers
-        )
-
-
 class ScanReportValueViewSetV2(viewsets.ModelViewSet):
     queryset = ScanReportValue.objects.order_by("id").only(
         "id", "value", "frequency", "value_description", "scan_report_field"
@@ -1170,19 +751,6 @@ class ScanReportValueViewSetV2(viewsets.ModelViewSet):
     @method_decorator(vary_on_cookie)
     def list(self, request, *args, **kwargs):
         return super().list(request, *args, **kwargs)
-
-
-class ScanReportValuesFilterViewSetScanReport(viewsets.ModelViewSet):
-    serializer_class = ScanReportValueViewSerializer
-    filter_backends = [DjangoFilterBackend]
-    filterset_fields = ["scan_report_field__scan_report_table__scan_report"]
-
-    def get_queryset(self):
-        return ScanReportValue.objects.filter(
-            scan_report_field__scan_report_table__scan_report=self.request.GET[
-                "scan_report"
-            ]
-        )
 
 
 class DownloadScanReportViewSet(viewsets.ViewSet):
